nephelios::services::helpers
Module docker_helper

Structs

AppInfo

AppMetadata

Functions

build_image

Builds a Docker image using the tarball created from the application directory.

check_swarm

Checks if Docker Swarm is active.

create_docker_context 🔒

Creates a Docker context tarball for the specified application path.

deploy_nephelios_stack

Runs the Docker Compose command to deploy the application. Creates and runs a Docker container from the specified image.

generate_and_write_dockerfile

Generates and writes a Dockerfile for the given application type.

get_app_status

init_swarm

Initializes Docker Swarm with the given IP address.

is_app_running 🔒

leave_swarm

Leaves the Docker Swarm.

list_deployed_apps

prune_images

Prunes unused Docker images.

push_image

Pushes a Docker image to a remote registry.

remove_service

Removes the container for the given application.

scale_app

Scales a Docker service.

stop_nephelios_stack

Stops the Nephelios stack by removing the Docker stack.