nephelios::services::helpers::docker_helper

Function get_app_status

pub async fn get_app_status(name: String) -> String

Retrieves the status of the specified application.

This function checks if the application is running, stopping, or stopped by using helper functions is_app_running and is_app_stopping.

Arguments

name - The name of the application whose status is to be retrieved.

Returns

A String representing the status of the application:

"running" if the application is currently running.

"stopping" if the application is in the process of stopping.

"stopped" if the application is not running.

"unknown" if the status could not be determined.